requirements.txt
python-telegram-bot[job-queue]
requests
pytz
flask
pymongo
dnspython
